requirements.txt
black
isort
mypy
pydocstyle
pylint
pytest
setuptools
tqdm
wheel
scipy
opencv-python
einops
matplotlib
timm
fvcore
PyWavelets
scikit-image
tensorboardX
thop
torchinfo
pandas
torchmetrics
